requirements.txt
flask==3.0.3
loguru==0.7.2
requests==2.32.3
torch==2.3.1+cu121
pyyaml==6.0.1
transformers==4.43.3
dataclasses_json==0.6.7
sentencepiece==0.2.0
pygame==2.6.0
pyaudio==0.2.14
keyboard==0.13.5
selenium==4.23.1
pyautogui==0.9.54
bilibili-api-python==16.2.0
PyQt5==5.15.11
scipy==1.14.0